# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is the API documentation for the Dora Explorer application.",
        "title": "Dora Explorer API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/v1/epoch/{epoch}": {
            "get": {
                "description": "Returns information for a specified epoch by the epoch number or an epoch tag (can be latest or finalized). On round-based networks, epoch_votes_unavailable is true and globalparticipationrate and votedether are null; use round participation instead.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Epoch"
                ],
                "summary": "Get epoch by number, latest, finalized",
                "operationId": "getEpoch",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Epoch number, the string latest or the string finalized",
                        "name": "epoch",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/api.APIEpochResponseV1"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Failure",
                        "schema": {
                            "$ref": "#/definitions/api.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ApiResponse"
                        }
                    }
                }
            }
        },
...
        "/v1/epochs": {
            "get": {
                "description": "Returns a list of epochs with detailed information and statistics. On round-based networks, epoch_votes_unavailable is true and target_voted, head_voted, total_voted, and vote_participation are null; use round participation instead.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "epochs"
                ],
                "summary": "Get epochs list",
                "operationId": "getEpochs",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "First epoch to display",
                        "name": "epoch",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of epochs to return (max 100)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.APIEpochsResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid parameters",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.APIEpochInfo": {
            "type": "object",
            "properties": {
                "attestations": {
                    "type": "integer"
                },
                "attester_slashings": {
                    "type": "integer"
                },
                "bls_changes": {
                    "type": "integer"
                },
                "consolidation_requests": {
                    "type": "integer"
                },
                "deposits": {
                    "type": "integer"
                },
                "deposits_amount": {
                    "type": "integer"
                },
                "eligible_ether": {
                    "type": "integer"
                },
                "epoch": {
                    "type": "integer"
                },
                "epoch_votes_unavailable": {
                    "type": "boolean"
                },
                "exits": {
                    "type": "integer"
                },
                "finalized": {
                    "type": "boolean"
                },
                "head_voted": {
                    "type": "integer",
                    "x-nullable": true
                },
                "max_sync_committee_size": {
                    "type": "integer"
                },
                "min_sync_committee_size": {
                    "type": "integer"
                },
                "missed_blocks": {
                    "type": "integer"
                },
                "missed_payloads": {
                    "type": "integer"
                },
                "orphaned_blocks": {
                    "type": "integer"
                },
                "proposed_blocks": {
                    "type": "integer"
                },
                "proposed_payloads": {
                    "type": "integer"
                },
                "proposer_slashings": {
                    "type": "integer"
                },
                "sync_participation": {
                    "type": "number"
                },
                "target_voted": {
                    "type": "integer",
                    "x-nullable": true
                },
                "total_voted": {
                    "type": "integer",
                    "x-nullable": true
                },
                "validator_balance": {
                    "type": "integer"
                },
                "validators": {
                    "type": "integer"
                },
                "vote_participation": {
                    "type": "number",
                    "x-nullable": true
                },
                "voting_finalized": {
                    "type": "boolean"
                },
                "voting_justified": {
                    "type": "boolean"
                },
                "withdrawal_requests": {
                    "type": "integer"
                },
                "withdrawals_amount": {
                    "type": "integer"
                },
                "withdrawals_count": {
                    "type": "integer"
                }
            }
        },
        "api.APIEpochResponseV1": {
            "type": "object",
            "properties": {
                "attestationscount": {
                    "type": "integer"
                },
                "attesterslashingscount": {
                    "type": "integer"
                },
                "averagevalidatorbalance": {
                    "type": "integer"
                },
                "blockscount": {
                    "type": "integer"
                },
                "depositscount": {
                    "type": "integer"
                },
                "eligibleether": {
                    "type": "integer"
                },
                "epoch": {
                    "type": "integer"
                },
                "epoch_votes_unavailable": {
                    "type": "boolean"
                },
                "finalized": {
                    "type": "boolean"
                },
                "globalparticipationrate": {
                    "type": "integer",
                    "x-nullable": true
                },
                "missedblocks": {
                    "type": "integer"
                },
                "missedpayloads": {
                    "type": "integer"
                },
                "orphanedblocks": {
                    "type": "integer"
                },
                "proposedblocks": {
                    "type": "integer"
                },
                "proposedpayloads": {
                    "type": "integer"
                },
                "proposerslashingscount": {
                    "type": "integer"
                },
                "rewards_exported": {
                    "type": "integer"
                },
                "scheduledblocks": {
                    "type": "integer"
                },
                "totalvalidatorbalance": {
                    "type": "integer"
                },
                "ts": {
                    "type": "integer"
                },
                "validatorscount": {
                    "type": "integer"
                },
                "voluntaryexitscount": {
                    "type": "integer"
                },
                "votedether": {
                    "type": "integer",
                    "x-nullable": true
                },
                "withdrawalcount": {
                    "type": "integer"
                }
            }
        },
        "api.APIEpochsData": {
            "type": "object",
            "properties": {
                "current_epoch": {
                    "type": "integer"
                },
                "epoch_count": {
                    "type": "integer"
                },
                "epochs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.APIEpochInfo"
                    }
                },
                "finalized_epoch": {
                    "type": "integer"
                },
                "first_epoch": {
                    "type": "integer"
                },
                "last_epoch": {
                    "type": "integer"
                },
                "total_epochs": {
                    "type": "integer"
                }
            }
        },
        "api.APIEpochsResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/api.APIEpochsData"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "api.ApiResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "status": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}