# handlers/api/epoch_v1.go
package api

import (
	"encoding/json"
	"fmt"
	"net/http"
	"strconv"

	"github.com/ethpandaops/dora/services"
	"github.com/ethpandaops/dora/types/models"
	"github.com/ethpandaops/go-eth2-client/spec/phase0"
	"github.com/gorilla/mux"
	"github.com/sirupsen/logrus"
)

type APIEpochResponseV1 struct {
	EpochVotesUnavailable   bool   `json:"epoch_votes_unavailable"`
	Epoch                   uint64 `json:"epoch"`
	Ts                      uint64 `json:"ts"`
	AttestationsCount       uint64 `json:"attestationscount"`
	AttesterSlashingsCount  uint64 `json:"attesterslashingscount"`
	AverageValidatorBalance uint64 `json:"averagevalidatorbalance"`
	BlocksCount             uint64 `json:"blockscount"`
	DepositsCount           uint64 `json:"depositscount"`
	EligibleEther           uint64 `json:"eligibleether"`
	Finalized               bool   `json:"finalized"`
	GlobalParticipationRate uint64 `json:"globalparticipationrate" extensions:"x-nullable"`
	MissedBlocks            uint64 `json:"missedblocks"`
	OrphanedBlocks          uint64 `json:"orphanedblocks"`
	ProposedBlocks          uint64 `json:"proposedblocks"`
	ProposedPayloads        uint64 `json:"proposedpayloads"`
	MissedPayloads          uint64 `json:"missedpayloads"`
	ProposerSlashingsCount  uint64 `json:"proposerslashingscount"`
	ScheduledBlocks         uint64 `json:"scheduledblocks"`
	TotalValidatorBalance   uint64 `json:"totalvalidatorbalance"`
	ValidatorsCount         uint64 `json:"validatorscount"`
	VoluntaryExitsCount     uint64 `json:"voluntaryexitscount"`
	VotedEther              uint64 `json:"votedether" extensions:"x-nullable"`
	RewardsExported         uint64 `json:"rewards_exported"`
	WithdrawalCount         uint64 `json:"withdrawalcount"`
}

func (data *APIEpochResponseV1) MarshalJSON() ([]byte, error) {
	type plain APIEpochResponseV1
	return models.MarshalEpochVotes((*plain)(data), data.EpochVotesUnavailable, "globalparticipationrate", "votedether")
}

// ApiEpoch godoc
// @Summary Get epoch by number, latest, finalized
// @Tags Epoch
// @Description Returns information for a specified epoch by the epoch number or an epoch tag (can be latest or finalized). On round-based networks, epoch_votes_unavailable is true and globalparticipationrate and votedether are null; use round participation instead.
// @Produce  json
// @Param  epoch path string true "Epoch number, the string latest or the string finalized"
// @Success 200 {object} ApiResponse{data=APIEpochResponseV1} "Success"
// @Failure 400 {object} ApiResponse "Failure"
// @Failure 500 {object} ApiResponse "Server Error"
// @Router /v1/epoch/{epoch} [get]
// @ID getEpoch
func ApiEpochV1(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json")

	vars := mux.Vars(r)

	epoch, err := strconv.ParseInt(vars["epoch"], 10, 64)
	if err != nil && vars["epoch"] != "latest" && vars["epoch"] != "finalized" {
		sendBadRequestResponse(w, r.URL.String(), "invalid epoch provided")
		return
	}

	chainState := services.GlobalBeaconService.GetChainState()
	if vars["epoch"] == "latest" {
		epoch = int64(chainState.CurrentEpoch())
	}

	finalizedEpoch, _ := chainState.GetFinalizedCheckpoint()
	if vars["epoch"] == "finalized" {
		epoch = int64(finalizedEpoch)
	}

	if epoch > int64(chainState.CurrentEpoch()) {
		sendBadRequestResponse(w, r.URL.String(), fmt.Sprintf("epoch is in the future. The latest epoch is %v", chainState.CurrentEpoch()))
		return
	}

	if epoch < 0 {
		sendBadRequestResponse(w, r.URL.String(), "epoch must be a positive number")
		return
	}

	data := &APIEpochResponseV1{
		EpochVotesUnavailable: chainState.SlotsPerRound() > 0,
		Epoch:                 uint64(epoch),
		Ts:                    uint64(chainState.EpochToTime(phase0.Epoch(epoch)).Unix()),
		Finalized:             chainState.IsEpochFinalized(phase0.Epoch(epoch)),
	}

	dbEpochs := services.GlobalBeaconService.GetDbEpochs(r.Context(), uint64(epoch), 1)
	dbEpoch := dbEpochs[0]
	if dbEpoch != nil {
		data.AttestationsCount = dbEpoch.AttestationCount
		data.DepositsCount = dbEpoch.DepositCount
		data.VoluntaryExitsCount = dbEpoch.ExitCount
		data.WithdrawalCount = dbEpoch.WithdrawCount
		data.RewardsExported = dbEpoch.WithdrawAmount
		data.ProposerSlashingsCount = dbEpoch.ProposerSlashingCount
		data.AttesterSlashingsCount = dbEpoch.AttesterSlashingCount
		data.EligibleEther = dbEpoch.Eligible
		data.VotedEther = dbEpoch.VotedTotal
		if dbEpoch.Eligible > 0 {
			data.GlobalParticipationRate = uint64(dbEpoch.VotedTarget * 100 / dbEpoch.Eligible)
		}
		data.ValidatorsCount = dbEpoch.ValidatorCount
		data.TotalValidatorBalance = dbEpoch.ValidatorBalance
		if dbEpoch.ValidatorCount > 0 {
			data.AverageValidatorBalance = dbEpoch.ValidatorBalance / dbEpoch.ValidatorCount
		}

		specs := chainState.GetSpecs()
		currentSlot := chainState.CurrentSlot()
		lastSlot := chainState.EpochToSlot(phase0.Epoch(epoch+1)) - 1
		passedSlotCount := uint64(specs.SlotsPerEpoch)
		if currentSlot < lastSlot {
			data.ScheduledBlocks = passedSlotCount - uint64(lastSlot-currentSlot)
			passedSlotCount -= data.ScheduledBlocks
		}

		data.MissedBlocks = passedSlotCount - uint64(dbEpoch.BlockCount)
		data.BlocksCount = uint64(dbEpoch.BlockCount + dbEpoch.OrphanedCount)
		data.OrphanedBlocks = uint64(dbEpoch.OrphanedCount)
		data.ProposedBlocks = uint64(dbEpoch.BlockCount)

		// Pre-ePBS the execution payload is bundled inside the beacon block, so every
		// canonical block carries a payload. Post-ePBS (EIP-7732) payloads are revealed
		// separately and may be missing, so use the dedicated payload count.
		proposedPayloads := uint64(dbEpoch.BlockCount)
		if chainState.IsEip7732Enabled(phase0.Epoch(epoch)) {
			proposedPayloads = dbEpoch.PayloadCount
		}
		data.ProposedPayloads = proposedPayloads
		// MissedPayloads counts canonical blocks that were proposed but whose payload was
		// not revealed (an ePBS-specific failure, distinct from MissedBlocks). Always 0 pre-ePBS.
		data.MissedPayloads = uint64(dbEpoch.BlockCount) - proposedPayloads
	}

	j := json.NewEncoder(w)
	response := &ApiResponse{}
	response.Status = "OK"
	response.Data = data

	err = j.Encode(response)
	if err != nil {
		sendServerErrorResponse(w, r.URL.String(), "could not serialize data results")
		logrus.Errorf("error serializing json data for API %v route: %v", r.URL, err)
	}
}